+++ src/main/java/dev/yatloaf/modkrowd/config/feature/CherryLiteThemeFeature.java
@@ -55,6 +55,7 @@ public class CherryLiteThemeFeature extends ThemeFeature {
 
     @Override
     public void onMessage(MessageCache message, MinecraftClient client, ActionQueue queue) {
+        message.setBackgroundTint(STEM_T);
         switch (message.result()) {
             case AlohaMessage alohaMessage -> message.setThemed(this.alohaMessage(alohaMessage));
             case AfkMessage afkMessage -> message.setThemed(this.afkMessage(afkMessage));