+++ src/main/java/dev/yatloaf/modkrowd/mixinduck/ChatHudLineDuck.java
@@ -1,0 +1,6 @@
+package dev.yatloaf.modkrowd.mixinduck;
+
+public interface ChatHudLineDuck {
+    void modKrowd$setBackgroundTint(int color);
+    int modKrowd$getBackgroundTint();
+}